+++ internal/engine/docker.go
@@ -27,7 +27,7 @@ func newDocker(ctx context.Context) (*DockerEngine, error) {
 	return &DockerEngine{}, nil
 }
 
-func (c *DockerEngine) BuildImage(ctx context.Context, project files.ProjectEntry, relativeDotfilesDir string) error {
+func (c *DockerEngine) BuildImage(ctx context.Context, project files.ProjectEntry) error {
 	buildCfg, err := loadBuildConfig(project)
 	if err != nil {
 		return err
@@ -47,7 +47,6 @@ func (c *DockerEngine) BuildImage(ctx context.Context, project files.ProjectEntr
 	for _, key := range keys {
 		cmdArgs = append(cmdArgs, "--build-arg", fmt.Sprintf("%s=%s", key, buildCfg.Args[key]))
 	}
-	cmdArgs = append(cmdArgs, "--build-arg", "DOTFILES_DIR="+relativeDotfilesDir)
 	cmdArgs = append(cmdArgs, projectBaseDataDir(project))
 
 	cmd := exec.CommandContext(ctx, "docker", cmdArgs...)
@@ -92,6 +91,19 @@ func (c *DockerEngine) RunContainer(ctx context.Context, project files.ProjectEn
 		"--volume", runtimeCfg.ProjectPath + ":" + projectMount,
 		"--volume", "paulenv-shared-cache:/home/" + username + "/.container-cache",
 		"--volume", projectLocalVolumeName(project.ProjectName) + ":/home/" + username + "/.container-local",
+	}
+	if runtimeCfg.DotfilesPath != "" {
+		dotfilesPath, err := resolveRuntimePath(project.RuntimeConfigPath, runtimeCfg.DotfilesPath)
+		if err != nil {
+			return fmt.Errorf("resolve DOTFILES_PATH: %w", err)
+		}
+		cmdArgs = append(cmdArgs, "--volume", dotfilesPath+":/paul-env/dotfiles:ro")
+	}
+	if runtimeCfg.GitName != "" {
+		cmdArgs = append(cmdArgs, "--env", "GIT_AUTHOR_NAME="+runtimeCfg.GitName)
+	}
+	if runtimeCfg.GitEmail != "" {
+		cmdArgs = append(cmdArgs, "--env", "GIT_AUTHOR_EMAIL="+runtimeCfg.GitEmail)
 	}
 
 	for _, volume := range runtimeCfg.Volumes {